EXP-5
Draw the use case diagram is a graphic depiction of the interactions among the elements of the Railway reservation system for maintaining admin user can search ticket, view the description of a selected ticket, add a ticket, update a ticket and delete a ticket and its shows the activity flow of editing, adding and updating of the customer. The user will be able to search and generate reports of payment, Booking and train schedules.