EXP-3
Draw and validate Flowchart to compute quotient and remainder between two integers can be calculated by using the division ( / ) and modulus ( % ) operators respectively. To compute the remainder of the division of two floating point numbers, the library function fmod() is used. This function considers quotient as an integer number and the remainder as a floating point number