EXP-11
Draw a Class diagram for Online Voting System using CASE tools. eVote is an election system that facilitates voters to record their secure and secret ballot electronically. It has a friendly user interface and enables voters to cast their votes in few simple steps. We ensures the authenticity of the voters and the votes cast by them along with non-traceability of the casted vote elaborate the whole process of e-voting