EXP-12
Draw a Class diagram for Library Management System using CASE tools.designing use case diagram for the library management system. User who registers himself as a new user initially is regarded as staff or student for the library system. Librarian has a key role in this system. Librarian adds the records in the library database about each student or user every time issuing the book or returning the book, or paying fine. Design the class diagram for this