EXP-13
Draw a Activity diagram for Online Voting System for a software platform that allows groups to securely conduct votes and elections using CASE tools. The voters should be able to register via some proper authority. Construct a system that the voters should see the list of candidates present in his constituency. A voter should be able to cast his vote to a candidate and that voter should cast only 1 vote at a timeDraw a Activity diagram for Library Management System using CASE tools.