EXP-14
Draw a Activity diagram for Library Management System using CASE tools constructing User who registers himself as a new user initially who is regarded as staff or student. After getting the library card, a new book is requested by the user. After, requesting, the desired book issue the book to the user. Renew the book if the user needs the book again. Add the agent as Librarian and the user in the activity diagram